+++ src/ed_crystal_coin/transaction.cr
@@ -1,7 +1,7 @@
 class EdCrystalCoin::Block::Transaction
   getter from : String
   getter to : String
-  getter amount : Int32
+  getter amount : Int64
 
   def initialize(@from, @to, @amount); end
 end
